Optimizer: convert the AliasInfoDumper and  MemBehaviorDumper passes to tests
