+++ test/SILOptimizer/alias-analysis_unit.sil
@@ -1,4 +1,4 @@
-// RUN: %target-sil-opt %s -dump-alias-info -o /dev/null | %FileCheck %s
+// RUN: %target-sil-opt %s -test-runner -o /dev/null | %FileCheck %s
 
 // General black-box alias analysis tests. White-box unit tests are
 // specific to the aa-kind, such as basic-aa.sil and typed-access-tb-aa.sil.
@@ -28,6 +28,7 @@ struct OuterStruct {
 // CHECK-NEXT:    NoAlias
 sil shared @testOffsetBehindProjection : $@convention(thin) (@inout MyStruct) -> () {
 bb0(%0 : $*MyStruct):
+  specify_test "aliasing"
   %1 = struct_element_addr %0 : $*MyStruct, #MyStruct.i
   %2 = integer_literal $Builtin.Word, 1
   %3 = index_addr %1 : $*Int64, %2 : $Builtin.Word
@@ -46,6 +47,7 @@ bb0(%0 : $*MyStruct):
 // CHECK-NEXT:    NoAlias
 sil shared @testOffsetsBehindProjectionOverlap : $@convention(thin) (@inout OuterStruct) -> () {
 bb0(%0 : $*OuterStruct):
+  specify_test "aliasing"
   %1 = struct_element_addr %0 : $*OuterStruct, #OuterStruct.s
   %2 = struct_element_addr %1 : $*MyStruct, #MyStruct.i
   %3 = load %2 : $*Int64
@@ -70,6 +72,7 @@ bb0(%0 : $*OuterStruct):
 // CHECK-NEXT:   MayAlias
 sil @testIndexOf0 : $@convention(thin) (@inout MyStruct) -> () {
 bb0(%0 : $*MyStruct):
+  specify_test "aliasing"
   %1 = struct_element_addr %0 : $*MyStruct, #MyStruct.i
   %2 = integer_literal $Builtin.Word, 0
   %3 = index_addr %0 : $*MyStruct, %2 : $Builtin.Word
@@ -91,6 +94,7 @@ bb0(%0 : $*MyStruct):
 // CHECK-NEXT:    MayAlias
 sil @testUnknownIndex : $@convention(thin) (@inout MyStruct, Builtin.Word) -> () {
 bb0(%0 : $*MyStruct, %1 : $Builtin.Word):
+  specify_test "aliasing"
   %2 = struct_element_addr %0 : $*MyStruct, #MyStruct.i
   %3 = index_addr %0 : $*MyStruct, %1 : $Builtin.Word
   %4 = struct_element_addr %3 : $*MyStruct, #MyStruct.i
@@ -127,6 +131,7 @@ bb0(%0 : $*MyStruct, %1 : $Builtin.Word):
 // CHECK-NEXT:    NoAlias
 sil @testVectorBaseAddr : $@convention(thin) (@inout Builtin.FixedArray<10, Int>, Builtin.Word) -> () {
 bb0(%0 : $*Builtin.FixedArray<10, Int>, %1 : $Builtin.Word):
+  specify_test "aliasing"
   %2 = vector_base_addr %0
   %3 = index_addr %2, %1
   %4 = integer_literal $Builtin.Word, 1
