+++ test/SILOptimizer/typed-access-tb-aa_unit.sil
@@ -1,4 +1,4 @@
-// RUN: %target-sil-opt %s -dump-alias-info -o /dev/null | %FileCheck %s
+// RUN: %target-sil-opt %s -test-runner -o /dev/null | %FileCheck %s
 
 // REQUIRES: asserts
 
@@ -39,6 +39,7 @@ sil @goo_init : $@convention(thin) (@thick goo.Type) -> @owned goo
 // CHECK-NEXT: NoAlias
 sil hidden @no_parent_child_relation_reftype_tests : $@convention(thin) () -> () {
 bb0:
+  specify_test "aliasing"
   %0 = alloc_stack $boo, var, name "a"                 // users: %5, %14
   %1 = alloc_stack $baz, var, name "b"                 // users: %9, %13
   %2 = function_ref @boo_init : $@convention(thin) (@thick boo.Type) -> @owned boo // user: %4
@@ -65,6 +66,7 @@ bb0:
 // CHECK-NEXT: MayAlias
 sil hidden @with_parent_child_relation_reftype_tests : $@convention(thin) () -> () {
 bb0:
+  specify_test "aliasing"
   %0 = alloc_stack $boo, var, name "a"                 // users: %5, %14
   %1 = alloc_stack $goo, var, name "b"                 // users: %9, %13
   %2 = function_ref @boo_init : $@convention(thin) (@thick boo.Type) -> @owned boo // user: %4
@@ -210,6 +212,7 @@ bb0:
 
 sil @builtin_test : $@convention(thin) () -> () {
 bb0:
+  specify_test "aliasing"
   %0 = pointer_to_address undef : $Builtin.RawPointer to [strict] $*Builtin.RawPointer
   %1 = pointer_to_address undef : $Builtin.RawPointer to [strict] $*Builtin.NativeObject
   %2 = pointer_to_address undef : $Builtin.RawPointer to [strict] $*Builtin.AnyObject
@@ -321,6 +324,7 @@ struct STest_S3 {
 
 sil @struct_tests : $@convention(thin) () -> () {
 bb0:
+  specify_test "aliasing"
   %0 = pointer_to_address undef : $Builtin.RawPointer to [strict] $*STest_S1
   %1 = pointer_to_address undef : $Builtin.RawPointer to [strict] $*STest_S2
   %2 = pointer_to_address undef : $Builtin.RawPointer to [strict] $*STest_E1
@@ -436,6 +440,7 @@ enum ETest_E3 {
 
 sil @enum_tests : $@convention(thin) () -> () {
 bb0:
+  specify_test "aliasing"
   %0 = pointer_to_address undef : $Builtin.RawPointer to [strict] $*ETest_S1
   %1 = pointer_to_address undef : $Builtin.RawPointer to [strict] $*ETest_S2
   %2 = pointer_to_address undef : $Builtin.RawPointer to [strict] $*ETest_E1
@@ -526,6 +531,7 @@ protocol AnyObject {}
 
 sil @class_tests : $@convention(thin) () -> () {
 bb0:
+  specify_test "aliasing"
   %0 = pointer_to_address undef : $Builtin.RawPointer to [strict] $*CTest_C1
   %1 = pointer_to_address undef : $Builtin.RawPointer to [strict] $*CTest_C2
   %2 = pointer_to_address undef : $Builtin.RawPointer to [strict] $*CTest_C3
@@ -628,6 +634,7 @@ enum TTest_E1 {
 
 sil @tuple_tests : $@convention(thin) () -> () {
 bb0:
+  specify_test "aliasing"
   %0 = pointer_to_address undef : $Builtin.RawPointer to [strict] $*(Builtin.RawPointer, Builtin.Int64)
   %1 = pointer_to_address undef : $Builtin.RawPointer to [strict] $*(TTest_S1, Builtin.Int64)
   %2 = pointer_to_address undef : $Builtin.RawPointer to [strict] $*(TTest_E1, Builtin.Int64)
@@ -666,6 +673,7 @@ bb0:
 // CHECK-NEXT: MayAlias
 sil @tbaa_dump : $@convention(thin) (@in Builtin.Int64) -> Builtin.Int64 {
 bb0(%0 : $*Builtin.Int64):
+  specify_test "aliasing"
   %1 = integer_literal $Builtin.Int64, 42
   store %1 to %0 : $*Builtin.Int64
   %3 = unchecked_addr_cast %0 : $*Builtin.Int64 to $*Builtin.Int32
@@ -687,6 +695,7 @@ bb0(%0 : $*Builtin.Int64):
 // CHECK-NEXT: NoAlias
 sil @test_project_box : $@convention(thin) (<τ_0_0> { var τ_0_0 } <Builtin.Int32>, <τ_0_0> { var τ_0_0 } <Builtin.Int64>) -> () {
 bb0(%0 : $<τ_0_0> { var τ_0_0 } <Builtin.Int32>, %1 : $<τ_0_0> { var τ_0_0 } <Builtin.Int64>):
+  specify_test "aliasing"
   %2 = project_box %0 : $<τ_0_0> { var τ_0_0 } <Builtin.Int32>, 0
   %3 = project_box %1 : $<τ_0_0> { var τ_0_0 } <Builtin.Int64>, 0
 
@@ -704,6 +713,7 @@ bb0(%0 : $<τ_0_0> { var τ_0_0 } <Builtin.Int32>, %1 : $<τ_0_0> { var τ_0_0 }
 // CHECK-NEXT: NoAlias
 sil @test_ref_tail_addr : $@convention(thin) (CTest_C1, CTest_C1) -> () {
 bb0(%0 : $CTest_C1, %1 : $CTest_C1):
+  specify_test "aliasing"
   %2 = ref_element_addr %0 : $CTest_C1, #CTest_C1.x
   %3 = ref_tail_addr %1 : $CTest_C1, $Builtin.Int16
 
@@ -726,6 +736,7 @@ struct GenericStruct<T> {
 
 sil @test_generic_struct : $@convention(thin) <T> (@thick T.Type) -> () {
 bb0(%0 : $@thick T.Type):
+  specify_test "aliasing"
   %1 = pointer_to_address undef : $Builtin.RawPointer to [strict] $*Builtin.Int8
   %2 = pointer_to_address undef : $Builtin.RawPointer to [strict] $*GenericStruct<T>
